ci: use ansible-community/ansible-test-gh-action action
This is an already-made action made by the Ansible community that runs
ansible-test with extra bits on top of it; using it allows us to get
rid of the manual installation of Python and Ansible, automatically done
by the action.

One big gotcha is the lack of galaxy.yml in the repository (generated
during release time, see also #19); to overcome this, we need some more
steps:
- keep checking out the repository manually (ansible-test-gh-action is
  able to do it automatically), and fetch all the history (needed for
  the later steps)
- checkout the target branch in the same way done by
  ansible-test-gh-action, matching the expectations of ansible-test
- generate galaxy.yml locally using the cuchi/jinja2-action action

Also, when using ansible < 2.12, pull it from felixfontein's fork, as it
contains the old "stable-N" branches.