fix: gracefully handle the lack of python-requests
Do not assume that requests is available, as the importing error will
not be properly noticeable in most of the environments; this is also
checked by the sanity testing of Ansible [1].

Follow the recommended practice by storing the import error separately,
raising it as Ansible error when the plugin is used. This makes it
possible to drop all the ignores for this issue for the sanity test.

[1] https://docs.ansible.com/ansible/latest/dev_guide/testing/sanity/import.html