ci: use own script for rendering a jinja2 template
It looks like the cuchi/jinja2-action action is currently broken [1].
Since all we need is to render a jinja2 template with string variables,
then roll out a simple script Python using the "jinja2" module to do
that.

The script adds to the render context all the environment variables
starting with the prefix "JINJA_", removing it.

The script does not have a ".py" extension on purpose, to avoid being
picked by ansible-test.

[1] https://github.com/cuchi/jinja2-action/issues/17

Signed-off-by: Pino Toscano <ptoscano@redhat.com>