chore: drop dead molecule setup
It has not been updated/touched after it was added with the role, more
than 3 years ago. Since it requires a special setup (Vagrant), and the
role is not going to get a lot more of work anyway, simply drop the bits
for molecule. This will simplify linting/checking the rest of the
codebase.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>