fix: set a bash shebang recognized by ansible-test
Few years ago, the "#!/bin/bash" shebang for bash scripts stopped to be
allowed by ansible-test in favour of "#!/usr/bin/env bash" [1].

Hence, switch the fact script to "#!/usr/bin/env bash", which will
behave in the same way. This makes it possible to drop the ignores for
this failure.

[1] https://github.com/ansible/ansible/pull/50954

Signed-off-by: Pino Toscano <ptoscano@redhat.com>