ci: simplify checkout for ansible-test
The "ansible-test-gh-action" action is invoked with a pre-existing
source directory (the "collection-src-directory" argument): this means
that the action will copy that directory, and create the directory
layout ("ansible_collections/<namespace>/<name>") automatically.
Because of that, there is no need to create that directory layout
manually when checking out the sources of the collection.

Hence:
- keep a subdirectory for the sources, so there is less chance to
  pollute them by writing directly to the workspace
- shorten the subdirectory of the sources to a simple "repo"

Signed-off-by: Pino Toscano <ptoscano@redhat.com>