ci: add ansible-lint workflow
Add a simple workflow to run ansible-lint on push and for pull requests.

ansible-lint requires a "ansible_collections/<namespace>/<name>/"
structure to recognize the collection and its bits, so check out the
code in that way. Also export few environment variables to locate all
the bits.

Sadly currently (ansible-lint v6.22.0) the action does not respect its
"working_directory" for all the steps, failing in its initial setup
steps; a potential fix has been proposed [1], so create a local
workaround to get the action to run in the meanwhile.

[1] https://github.com/ansible/ansible-lint/pull/3905

Signed-off-by: Pino Toscano <ptoscano@redhat.com>