ci: drop workaround for ansible-lint action bug
There was a bug [1] in the ansible-lint action that made its
"working_directory" parameter not working as expected. This was fixed in
v6.22.1 [2].

Since we use the "v6" tag, which points to the latest version of
ansible-lint 6.x, the fixed action should be used already. Hence, drop
the workaround step.

[1] https://github.com/ansible/ansible-lint/pull/3905
[2] https://github.com/ansible/ansible-lint/releases/tag/v6.22.1

Signed-off-by: Pino Toscano <ptoscano@redhat.com>