doc: improve example of inventory plugin
Having multiple examples for an inventory plugin in a single "example"
block is not seen nicely by ansible-lint: using an inventory plugin
means top-level YAML keys, so multiple examples will trigger duplicate
YAML keys.

Thus tweak the example to be a single example, with different sections
for the various parts; add a small section for the authentication.

The changes make it possible to drop the override for the ansible-lint
issue.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>