inventory: fix default of "staleness" option
Explicitly setting values means they will be passed in queries to
Inventory, filtering the results. The problem is that, in case new types
of staleness are added to Inventory, they will not be returned by
default, whereas default behaviour is to return the hosts with all the
types of staleness.

Since:
- Inventory does not do any filter if no staleness query arguments are
  passed
- an empty list of values for "staleness" already implies none are
  passed in the requests to Inventory
then change the default value of the "staleness" option to the empty
list.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>