compliance: manually check insights-client status
Manually check the Insights registration status using 'insights-client',
rather than relying on a fact exported by the 'insights_client' role.
That fact alone does not guarantee that the system is registered, as the
'machine-id' file may be on the system even after failed registration
attempts.

Bonus point: this makes it possible to use the 'compliance' role without
the 'insights_client' role, as long as something else takes care of
registering the system using 'insights-client'. Mention the 'rhc' role
as potential alternative.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>