inventory: manually check for required authentication options
Right now the "user" and "password" options are marked as mandatory, as
they are used for basic authentication which is currently the only
authentication method supported.

To prepare for additional authentication methods, then the two option
must not be marked as required anymore, and we rather need to check for
the options during the execution of the plugin.

There should be no behaviour change; the only change is what errors out
in case of missing options: now it is the plugin itself, rather than
Ansible.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>