inventory: bump patch hosts to 100
The default value of hosts per requests returned by Patchman is 20,
with a current maximum limit of 100. Hence, bump the number of hosts
returned per request to 100, to get the same details with one fifth
the amount of requests.

There should be no behaviour change, as pagination is already handled.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>