inventory: drop redundant "status_code == 200" checks
These code blocks had a pattern such as:

    if response.status_code != 200:
        raise AnsibleError("http error (%s): %s" %
                           (response.status_code, response.text))
    elif response.status_code == 200:
        # do something

The "status_code == 200" check is always true, since the check before
that was for the very opposite situation. Also, since an exception is
raisen when the status code is 200, no more code of the same function
will be run, and thus it is possible to deindent the whole block for the
status code 200.

This is a minor code refactoring with no behaviour changes.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>