inventory: fetching service account token again when expired
The access token got from the service account may have an expiration
time (in seconds); if the amount of data to request is big (e.g. there
are really many hosts in the Inventory), then the token may expire,
resulting in 401 errors.

The only way to avoid that is to fetch a new token when the existing one
is expired; hence:
- get the expiration time from the access token JSON, if available, and
  calculate the time in which the token will expire
- reduce the expiration of the token by 1 second: this will (greatly,
  hopefully) lower the chances of races between the validity check of
  the token and its usage
- when performing a new GET request, create a new bearer authentication
  fetching a new access token if the current expired; this will make
  sure that a new access token is requested only if needed

Signed-off-by: Pino Toscano <ptoscano@redhat.com>