feat: deprecate insights-client modules & plugins
Recent versions of the "rhc" Ansible system role, shipped as part of the
Linux System Roles collection, provide the same amount of Red Hat
Insights features as currently provided by the insights-client bits:
- the insights_config action/module
- the insights_register module
- the insights_client role

Since the "rhc" system role is the preferred solution for
registration/connection to Red Hat Insights, formally deprecate all the
insights-client bits, slating them for removal in the next major
version (2.0.0).

The deprecation messages are emitted for each of the deprecated bit, and
they are shown only if the deprecation messages are not explicitly
disabled by the user. Since the name of the collection is now present in
a couple of files, add them to the list of files in which update the
collection name on release.

[1] https://github.com/linux-system-roles/rhc

Signed-off-by: Pino Toscano <ptoscano@redhat.com>