fix(ci): ansible-test: set the collection namespace to "redhat"
The collection namespace set in the sources is "redhat" already, and it
is replaced with the proper one on release (or in general, when running
the release.yml playbook manually).

If the collection namespace is taken from the GitHub owner of the
repository for which the ansible-test workflow runs, then there will be
a mismatch between what is set in galaxy.yml and what is found in the
bits of the collection.

Hence, set the collection namespace to "redhat" when generating
galaxy.yml to create the right source tree for ansible-test.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>