ci: create local commit with generated galaxy.yml
ansible-test uses git underneath to detect the files of the collection;
since the autogenerated galaxy.yml file is not in the repository, then
it is not packed as part of the sources that are run in virtual
environments (e.g. docker, venv, etc).

Create a local commit with galaxy.yml so that it is properly accounted
for.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>