chore: simplify .gitignore
Clean it up a bit from things that are unrelated, leaving only the bits
for files actually generated by using/building this repository.

This cleanup also includes all the bits related to editors/IDEs: there
are way too many of them, and most likely almost none is actually used
by people working in this repository. If needed, it is easy to
selectively add needed bits.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>